Some users were reporting problems with the compat package where the shell exceeded the number of arguments possible. This was caused by the way were were handling CONFIG_* options -- we were just adding defines on the command line.
To correct this we now embrace our own autoconf header file, just as the kernel
has its own, with defines of the options specified in config.mk. Our header file
is self-generated by scripts/gen-compat-autoconf.sh which parses any
kernel-like config file and outputs a header file with proper defines.
The compat wireless autoconf file is include/linux/compat_autoconf.h and
should not be edited by hand.

The compat-wireless top Makefile makes uses of gen-compat-autoconf.sh through
scripts/check_config.sh. This script and the top level makefile
also check for updates to config.mk and if changes were made it forces
a regeneration for the compat autoconf file. To check for changes we are using
md5sum, which may be overkill, but should work well for now.

Additionally the top level makefile checks for the version of compat-wireless, if
it detects a new release has been pulled it will force a regeneration of the
compat autoconf as well.

This new build scheme been tested on 2.6.22 and 2.6.23.

External build has also been tested with 2.6.23:

make KLIB=/home/mcgrof/devel/kernels/linux-2.6.23.12 \
   KLIB_BUILD=/home/mcgrof/devel/kernels/linux-2.6.23.12

Signed-off-by: Luis R. Rodriguez <mcgrof@winlab.rutgers.edu>