Attached are two patches for compat. One just adds a .gitignore file so git status runs cleanly, and the other adds kernel 2.6.21 support and makes config.mk select drivers based on features like PCI, USB and the ITU-T CRC functions.
Note that I have only tested compiling everything but PCI drivers
against the old kernel so it may well be that some PCI functions are
missing in 2.6.21, they can be added then.

I had to disable rndis wlan because it uses cancel_work_sync() which is
only available in 2.6.22 and higher and can't be ported to older
kernels.

Feel free to reject the patch, with it, however, and a hack to allow
building the crypto code as a module, I now have mac80211 on my N810 :)
Not that I have a p54spi yet, but hey, it's a start.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
Signed-off-by: Luis R. Rodriguez <mcgrof@winlab.rutgers.edu>