Fork compat-wireless for old kernels <= 2.6.26
We fork compat-wireless for older kernels due to the new
TX multiqueue (MQ) changes. These new TX MQ changes are
painful to backport so instead of leaving everyone dry
we'll fork it, add new drivers as we can and start fresh
a new tree for >= 2.6.27.

This adds to git history drivers/ net/ and include/
with all the compat.diff changes applied. Note that
changes to compat.[ch] and mq_compat.h can now go directly
against their files based on the tree location, the compat/
directory is being removed as this is a complete fork
*with the tree*.

Patches for drivers can now be sent directly to this tree.

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>